% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/simdata_lasso_binomial.R
\docType{data}
\name{simdata_lasso_binomial}
\alias{simdata_lasso_binomial}
\title{Simulated complex survey data}
\format{
\subsection{\code{simdata_lasso_binomial}}{

A data frame with 1,720 rows and 54 columns:
\describe{
\item{y}{dichotomous response variable}
\item{x.1,...,x.50}{covariates}
\item{strata}{strata identifiers}
\item{cluster}{cluster identifiers}
\item{weights}{sampling weights}
}
}

\subsection{\code{simdata_lasso_binomial}}{

A data frame with 1720 rows and 54 columns:
\describe{
\item{x.1,...,x.50}{Covariates}
\item{y}{Dichotomous response variable}
\item{strata}{Column indicating the strata}
\item{cluster}{Column indicating the cluster}
\item{weights}{Sampling weights}
...
}
}
}
\usage{
simdata_lasso_binomial

simdata_lasso_binomial
}
\description{
This data has been simulated in order to provide the users with an example data set.
}
\keyword{datasets}
