% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welnet.plot.R
\name{welnet.plot}
\alias{welnet.plot}
\title{Plot welnet object}
\usage{
welnet.plot(x)
}
\arguments{
\item{x}{an object of class "welnet".}
}
\value{
a graph
}
\description{
Plot welnet object
}
\examples{
data(simdata_lasso_binomial)
mcv <- welnet(data = simdata_lasso_binomial, col.y = "y", col.x = 1:50,
              family = "binomial", alpha = 0.5, cluster = "cluster", strata = "strata",
              weights = "weights", method = "dCV", k=5, R=1)
welnet.plot(mcv)
}
